refactor(proxy): extract sse_stream spawn with cache hooks and bounded pre-generation retry
