+++ crates/gglib-proxy/src/lib.rs
@@ -13,6 +13,7 @@ pub mod models;
 pub mod server;
 pub mod slots;
 pub mod slots_poller;
+pub mod sse_stream;
 pub mod token_calibration;
 pub mod truncation;
 pub mod upstream_health;
